% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dir.R
\name{as_run_dir}
\alias{as_run_dir}
\title{Extract run directory from an object}
\usage{
as_run_dir(x)
}
\arguments{
\item{x}{Object to extract run directory from}
}
\value{
Run directory path(s)
}
\description{
Extract run directory from an object
}
\keyword{internal}
