% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_runs.R
\name{clean_runs}
\alias{clean_runs}
\title{Clean run directories}
\usage{
clean_runs(keep = NULL, project_dir = ".")
}
\arguments{
\item{keep}{Number of most recent runs to keep when cleaning runs.
\code{NULL} (the default) to remove all previous runs.}

\item{project_dir}{Project to list runs for}
}
\description{
Remove run directories from the filesystem.
}
\examples{
\dontrun{
clean_runs(keep = 10)
}
}
