% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_run.R
\name{compare_runs}
\alias{compare_runs}
\title{Compare training runs}
\usage{
compare_runs(runs = ls_runs(latest_n = 2),
  viewer = getOption("tfruns.viewer"))
}
\arguments{
\item{runs}{Character vector of 2 training run directories or
data frame returned from \code{\link[=ls_runs]{ls_runs()}} with at least 2 elements.}

\item{viewer}{Viewer to display training run information within
(default to an internal page viewer if available, otherwise
to the R session default web browser).}
}
\description{
Render a visual comparison of two training runs. The runs are
displayed with the most recent run on the right and the
earlier run on the left.
}
