% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy_run}
\alias{copy_run}
\alias{copy_run_files}
\alias{copy_runs}
\title{Copy run directories}
\usage{
copy_run(run_dir, to, overwrite = "confirm")

copy_run_files(run_dir, to, overwrite = "confirm")

copy_runs(run_dirs, to)
}
\arguments{
\item{run_dir}{Training run directory or data frame returned from
\code{\link[=ls_runs]{ls_runs()}}.}

\item{to}{Name of directory to copy into. For \code{copy_run} and \code{copy_run_files}
this is is a single directory which run files will be copied into. For
\code{copy_runs} this is a parent directory to copy multiple runs into as
subdirectories.}

\item{overwrite}{Whether to overwrite an existing \code{to} directory The default
is "confirm", which will promot in the case that \code{to} already exists.}

\item{run_dirs}{One or more run directories to copy into the
parent directory specified in \code{to}.}
}
\value{
Logical vector indicating which operation succeeded for each of the
run directories specified.
}
\description{
Functions for exporting/copying run directories and run artifact files.
}
\details{
Use \code{copy_run} to copy a single run directory to a named target
directory.

Use \code{copy_run_files} to copy only files saved/generated by a training run
script (e.g. saved models, checkpoints, etc.).

Use \code{copy_runs} to copy multiple runs into a parent directory.
}
\examples{
\dontrun{

# export a run directory
copy_run("runs/2017-09-24T10-54-00Z", to = "best-run")

# export the artifact files from a run directory
copy_run_files("runs/2017-09-24T10-54-00Z", "best-model")

# export the best (in terms of eval accuracy)
copy_runs(ls_runs(order = eval_acc)[1:3,], to = "best_eval_acc")

}
}
\seealso{
Other run management: \code{\link{clean_runs}}
}
