% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dir.R
\name{is_run_active}
\alias{is_run_active}
\title{Check for an active training run}
\usage{
is_run_active()
}
\value{
\code{TRUE} if a training tun is currently active
}
\description{
Check for an active training run
}
