% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_runs.R
\name{latest_run}
\alias{latest_run}
\alias{latest_runs}
\title{Enumerate recent training run directories}
\usage{
latest_run(project_dir = ".")

latest_runs(n, project_dir = ".")
}
\arguments{
\item{project_dir}{Project to list runs for}

\item{n}{Number of recent runs}
}
\value{
Character vector with run directory path(s)
}
\description{
Enumerate recent training run directories
}
