% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_runs.R
\name{ls_runs}
\alias{ls_runs}
\alias{latest_run}
\title{List training runs}
\usage{
ls_runs(subset = NULL, latest_n = NULL,
  runs_dir = getOption("tfruns.runs_dir", "runs"))

latest_run(runs_dir = getOption("tfruns.runs_dir", "runs"))
}
\arguments{
\item{subset}{Logical expression indicating rows to keep (missing values are
taken as false). See \code{\link[=subset]{subset()}}.}

\item{latest_n}{Limit query to the \code{latest_n} most recent runs}

\item{runs_dir}{Directory containing runs. Defaults to "runs" beneath the
current working directory (or to the value of the \code{tfruns.runs_dir} R
option if specified).}
}
\value{
Data frame with training run(s)
}
\description{
List training runs
}
