% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{run_dir}
\alias{run_dir}
\title{Training run directory}
\usage{
run_dir()
}
\value{
Current run direcotry (or \code{NULL} if no run directory is in use)
}
\description{
Returns the current run directory (if any)
}
\note{
You can also establish a run directory by defining the
\code{TENSORFLOW_RUN_DIR} environment variable (this is technically equivalent
to calling \code{\link[=initialize_run]{initialize_run()}} within an R script).
}
