% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_runs.R
\name{run_info}
\alias{run_info}
\title{Summary information for training runs}
\usage{
run_info(run_dir)
}
\arguments{
\item{run_dir}{Training run directory or data frame returned from
\code{\link[=ls_runs]{ls_runs()}}.}
}
\value{
Training run summary object with timing, flags, model info, training
and evaluation metrics, etc. If more than one \code{run_dir} is passed then
a list of training run summary objects is returned.
}
\description{
Summary information for training runs
}
