% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_run.R
\name{save_run_comparison}
\alias{save_run_comparison}
\title{Save a run comparison as HTML}
\usage{
save_run_comparison(runs = ls_runs(latest_n = 2), filename = "auto")
}
\arguments{
\item{runs}{Character vector of 2 training run directories or
data frame returned from \code{\link[=ls_runs]{ls_runs()}} with at least 2 elements.}

\item{filename}{Path to save the HTML to. If no \code{filename} is specified
then a temporary file is used (the path to the file is returned invisibly).}
}
\description{
Save a run comparison as HTML
}
