% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_run.R
\name{save_run_view}
\alias{save_run_view}
\title{Save a run view as HTML}
\usage{
save_run_view(run_dir = latest_run(), filename = "auto")
}
\arguments{
\item{run_dir}{Training run directory or data frame returned from
\code{\link[=ls_runs]{ls_runs()}}.}

\item{filename}{Path to save the HTML to. If no \code{filename} is specified
then a temporary file is used (the path to the file is returned invisibly).}
}
\description{
The saved view includes summary information (flags, metrics, model
attributes, etc.), plot and console output, and the code used for the run.
}
\seealso{
\code{\link[=ls_runs]{ls_runs()}}, \code{\link[=run_info]{run_info()}}, \code{\link[=view_run]{view_run()}}
}
