% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_run.R
\name{training_run}
\alias{training_run}
\title{Training run}
\usage{
training_run(path = ".", type = "local",
  config = Sys.getenv("R_CONFIG_ACTIVE", unset = "default"), flags = NULL,
  run_dir = NULL, echo = FALSE, envir = parent.frame(),
  encoding = getOption("encoding"))
}
\arguments{
\item{path}{Path to training script. Defaults to "train.R" within the current
working directory but can be any R script or directory that contains a
training script. If a directory is specified then it must either have a
file named "train.R" within it or otherwise have only a single R script
(failing that you should specify the full path to the R script).}

\item{type}{Run type (defaults to "local")}

\item{config}{The configuration to use. Defaults to the active configuration
for the current environment (as specified by the \code{R_CONFIG_ACTIVE}
environment variable), or \code{default} when unset.}

\item{flags}{Named character vector with flag values (see \code{\link[=flags]{flags()}}) or path
to YAML file containing flag values.}

\item{run_dir}{Directory to store run data within}

\item{echo}{Print expressions within training script}

\item{envir}{The environment in which the script should be evaluated}

\item{encoding}{The encoding of the training script; see \code{\link[=file]{file()}}.}
}
\value{
The directory used for the training run.
}
\description{
Run a training script with the specified \code{flags} within a unique run
directory.
}
