% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dir.R
\name{unique_run_dir}
\alias{unique_run_dir}
\title{Create a unique run directory}
\usage{
unique_run_dir(dir)
}
\arguments{
\item{dir}{Parent directory}
}
\description{
Create a new uniquely named run directory within the specified
parent \code{dir}.
}
\details{
The directory name will be a timestamp (in GMT time). If a
duplicate name is generated then the function will wait
long enough to return a unique one.
}
