% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_run.R
\name{view_run}
\alias{view_run}
\title{View a training run}
\usage{
view_run(run_dir = latest_run(), viewer = getOption("tfruns.viewer"))
}
\arguments{
\item{run_dir}{Training run directory or data frame returned from
\code{\link[=ls_runs]{ls_runs()}}.}

\item{viewer}{Viewer to display training run information within
(default to an internal page viewer if available, otherwise
to the R session default web browser).}
}
\description{
View metrics and other attributes of a training run.
}
\seealso{
\code{\link[=ls_runs]{ls_runs()}}, \code{\link[=run_info]{run_info()}}
}
