% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{write_run_data}
\alias{write_run_data}
\title{Write run metadata into the run directory}
\usage{
write_run_data(name, write_fn)
}
\arguments{
\item{name}{Name of metadata to write (subsequent writes with the same name
will overwrite).}

\item{write_fn}{Function that writes the data. The function will be
passed a single \code{data_dir} argument.}
}
\description{
Provide a function that will write data into the active run
directory (if any, no-op if there is none).
}
\seealso{
Other run data: \code{\link{write_run_metrics}}
}
