% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{write_run_metrics}
\alias{write_run_metrics}
\title{Write training run metrics}
\usage{
write_run_metrics(metrics)
}
\arguments{
\item{metrics}{Data frame containing run metrics.}
}
\description{
Write training metrics (e.g. accuracy, loss, etc.) into the active run
directory (if any, no-op if there is none).
}
\section{Metrics data format}{


Metrics should be passed as a data frame with one column for each metric.
If the metrics are not yet complete (e.g. only metrics for the
first several epochs are provided) then metrics in yet to be completed
epochs should use \code{NA} as their values. For example:\preformatted{data.frame':	30 obs. of  4 variables:
$ loss    : num  0.423 0.201 NA NA NA ...
$ acc     : num  0.873 0.942 NA NA NA ...
$ val_loss: num  0.174 0.121 NA NA NA ...
$ val_acc : num  0.949 0.964 NA NA NA ...
}
}

\seealso{
Other run data: \code{\link{write_run_data}}
}
