% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{clean_runs}
\alias{clean_runs}
\title{Clean run directories}
\usage{
clean_runs(runs_dir = "runs", keep = NULL)
}
\arguments{
\item{runs_dir}{Parent directory for runs}

\item{keep}{Number of most recent runs to keep when cleaning runs.
\code{NULL} (the default) to remove all previous runs.}
}
\description{
Remove run directories from the filesystem.
}
\examples{
\dontrun{
clean_runs(keep = 10)
}
}
