% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_runs.R
\name{clean_runs}
\alias{clean_runs}
\title{Clean run directories}
\usage{
clean_runs(runs = ls_runs(), keep = NULL, confirm = FALSE,
  runs_dir = getOption("tfruns.runs_dir", "runs"))
}
\arguments{
\item{runs}{Runs to clean}

\item{keep}{Number of most recent runs to keep when cleaning runs.
This will preserve the number of runs specified even if they
are provided in the list of \code{runs} to clean.}

\item{confirm}{\code{TRUE} to confirm before cleaning the runs.}

\item{runs_dir}{Directory containing runs. Defaults to "runs" beneath the
current working directory (or to the value of the \code{tfruns.runs_dir} R
option if specified).}
}
\description{
Remove run directories from the filesystem.
}
\examples{
\dontrun{
clean_runs(ls_runs(completed == FALSE))
clean_runs(keep = 10)
}
}
