% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy_run}
\alias{copy_run}
\alias{copy_run_files}
\alias{copy_runs}
\title{Copy run directories}
\usage{
copy_run(run_dir, to = ".", rename = NULL)

copy_run_files(run_dir, to = ".", rename = NULL)

copy_runs(run_dirs, to)
}
\arguments{
\item{run_dir}{Training run directory or data frame returned from
\code{\link[=ls_runs]{ls_runs()}}.}

\item{to}{Name of parent directory to copy run(s) into. Defaults to the
current working directory.}

\item{rename}{Rename run directory after copying. If not specified this
defaults to the basename of the run directory (e.g.
"2017-09-24T10-54-00Z").}

\item{run_dirs}{One or more run directories to copy into the
parent directory specified in \code{to}.}
}
\value{
Logical vector indicating which operation succeeded for each of the
run directories specified.
}
\description{
Functions for exporting/copying run directories and run artifact files.
}
\details{
Use \code{copy_run} to copy a single run directory

Use \code{copy_run_files} to copy only files saved/generated by a training run
script (e.g. saved models, checkpoints, etc.).

Use \code{copy_runs} to copy multiple run directories.
}
\examples{
\dontrun{

# export a run directory
copy_run("runs/2017-09-24T10-54-00Z", to = "exported-runs")

# export to the current working directory then rename
copy_run("runs/2017-09-24T10-54-00Z", rename = "best-run")

# export artifact files to the current working directory then rename
copy_run_files("runs/2017-09-24T10-54-00Z", rename = "best-model")

# export the best 3 runs (in terms of eval accuracy)
copy_runs(ls_runs(order = eval_acc)[1:3,], to = "best-runs")

}
}
\seealso{
Other run management: \code{\link{clean_runs}}
}
