% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{latest_run}
\alias{latest_run}
\alias{latest_runs}
\title{Enumerate recent training runs}
\usage{
latest_run(runs_dir = "runs")

latest_runs(runs_dir = "runs", n)
}
\arguments{
\item{runs_dir}{Parent directory for runs}

\item{n}{Number of recent runs}
}
\description{
Enumerate recent training runs
}
