% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_runs.R
\name{latest_run}
\alias{latest_run}
\alias{latest_runs}
\title{Enumerate recent training run directories}
\usage{
latest_run(runs_dir = getOption("tfruns.runs_dir", "runs"))

latest_runs(n, runs_dir = getOption("tfruns.runs_dir", "runs"))
}
\arguments{
\item{runs_dir}{Directory containing runs. Defaults to "runs"
beneath the current working directory (or to the value of the
\code{tfruns.runs_dir} R option if specified).}

\item{n}{Number of recent runs}
}
\value{
Character vector with run directory path(s)
}
\description{
Enumerate recent training run directories
}
