% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_runs.R
\name{list_runs}
\alias{list_runs}
\title{List training runs}
\usage{
list_runs(latest_n = NULL, project_dir = ".")
}
\arguments{
\item{latest_n}{Limit to a number of (most recent runs)}

\item{project_dir}{Project to list runs for}
}
\value{
A data frame with \code{created} (POSIXct) and \code{run_dir} (path relative to
\code{runs_dir}).
}
\description{
List training runs
}
