% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dir.R
\name{run_dir}
\alias{run_dir}
\title{Current run directory}
\usage{
run_dir()
}
\value{
Active run direcotry (or current working directory as a fallback)
}
\description{
Returns the current training run directory. If a training run is
not currently active (see \code{\link[=is_run_active]{is_run_active()}}) then the current
working directory is returned.
}
