% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_run.R
\name{training_run}
\alias{training_run}
\title{Training run}
\usage{
training_run(file = "train.R", runs_dir = "runs",
  config = Sys.getenv("R_CONFIG_ACTIVE", unset = "default"), flags = NULL,
  echo = FALSE, envir = parent.frame(), chdir = TRUE,
  encoding = getOption("encoding"))
}
\arguments{
\item{file}{Path to training script (defaults to "train.R")}

\item{runs_dir}{Directory to create run directories within}

\item{config}{The configuration to use. Defaults to the active configuration
for the current environment (as specified by the \code{R_CONFIG_ACTIVE}
environment variable), or \code{default} when unset.}

\item{flags}{Named character vector with flag values (see \code{\link[=flags]{flags()}})}

\item{echo}{Print expressions within training script}

\item{envir}{The environment in which the script should be evaluated}

\item{chdir}{If \code{TRUE} the R working directory is temporarily changed to the
directory containing file for evaluating.}

\item{encoding}{The encoding of the training script; see \code{\link[=file]{file()}}.}
}
\value{
The directory used for the training run.
}
\description{
Run a training script with the specified \code{flags} within a unique run directory.
}
