% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dir.R
\name{unique_run_dir}
\alias{unique_run_dir}
\title{Create a unique run directory}
\usage{
unique_run_dir(runs_dir = getOption("tfruns.runs_dir", "runs"))
}
\arguments{
\item{runs_dir}{Directory containing runs. Defaults to "runs" beneath the
current working directory (or to the value of the \code{tfruns.runs_dir} R
option if specified).}
}
\description{
Create a new uniquely named run directory within the specified
\code{runs_dir}.
}
\details{
The directory name will be a timestamp (in GMT time). If a
duplicate name is generated then the function will wait
long enough to return a unique one.
}
