% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_metadata.R
\name{write_run_data}
\alias{write_run_data}
\title{Write run data (deprecated)}
\usage{
write_run_data(type, data)
}
\arguments{
\item{type}{Type of metadata to write. Standard types include "flags",
"sources", "properties", "metrics", and "evaluation". You can
also specify a custom type (see \emph{Custom Types} section below).}

\item{data}{Metadata to write:
\itemize{
\item "flags" --- Named list of training flags
\item "source" --- Directory to copy source files from
\item "properties" --- Named list of arbitrary properties. Note
that properties will be stored as strings.
\item "metrics" --- Data frame with training run metrics
(see \emph{Metrics Data Frame} below).
\item "evaluation" --- Named list of evaluation metrics.
\item "error" --- Named list with 'message' and 'traceback'
\item "<custom>" -- Function used to write the data
(see \emph{Custom Types} section below).
}}
}
\description{
Deprecated alias for \code{\link[=write_run_metadata]{write_run_metadata()}}.
}
\keyword{internal}
